sports_framework/pyproject.toml
[project]
name = "sports-framework"
version = "0.1.0"
description = "Extensible sports betting ETL framework built on Dagster"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "Sports Analytics Team"}
]
keywords = ["sports", "betting", "etl", "dagster", "analytics"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    # Dagster Core
    "dagster>=1.7.0",
    "dagster-webserver>=1.7.0",
    "dagster-postgres>=0.22.0",
    
    # Data Processing
    "pandas>=2.1.0",
    "numpy>=1.26.0",
    "polars>=0.20.0",
    
    # API & HTTP
    "httpx>=0.26.0",
    "requests>=2.31.0",
    "aiohttp>=3.9.0",
    
    # Data Validation
    "pydantic>=2.5.0",
    "pandera>=0.18.0",
    
    # Database
    "sqlalchemy>=2.0.0",
    "psycopg[binary]>=3.1.0",
    "alembic>=1.13.0",
    
    # Caching
    "redis>=5.0.0",
    "cachetools>=5.3.0",
    
    # Utilities
    "python-dotenv>=1.0.0",
    "structlog>=24.1.0",
    "tenacity>=8.2.0",
    
    # Type hints
    "typing-extensions>=4.9.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.12.0",
    "mypy>=1.8.0",
    "ruff>=0.2.0",
    "black>=24.0.0",
    "pre-commit>=3.6.0",
]

ml = [
    "scikit-learn>=1.4.0",
    "xgboost>=2.0.0",
    "lightgbm>=4.1.0",
    "optuna>=3.5.0",
    "mlflow>=2.10.0",
]

viz = [
    "plotly>=5.18.0",
    "matplotlib>=3.8.0",
    "seaborn>=0.13.0",
    "streamlit>=1.30.0",
]

all = [
    "sports-framework[dev,ml,viz]",
]

[project.urls]
Homepage = "https://github.com/your-org/sports-framework"
Repository = "https://github.com/your-org/sports-framework"
Documentation = "https://sports-framework.readthedocs.io"
"Bug Tracker" = "https://github.com/your-org/sports-framework/issues"

[project.scripts]
sports-etl = "sports_framework.cli:main"

[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["sports_framework*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "--cov=sports_framework",
    "--cov-report=html",
    "--cov-report=term-missing",
    "--cov-fail-under=80",
]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true

[tool.ruff]
line-length = 100
target-version = "py310"
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
]
ignore = [
    "E501",  # line too long, handled by black
    "B008",  # do not perform function calls in argument defaults
]

[tool.black]
line-length = 100
target-version = ['py310']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''